Run set_date only if end_date is not defined in Email Campaign
set_date() is not allowing me to save any lead campaigns started before today. Even if I want to change the status or sender, that's not feasible for older records. If we put a condition to ensure set_date runs only if end_date is not defined, that solves this problem.